+++ src/main/java/ru/mephi/bot/cache/UserDataCache.java
@@ -2,7 +2,7 @@
 package ru.mephi.bot.cache;
 
 import org.springframework.stereotype.Component;
-import ru.mephi.config.bot.BotState;
+import ru.mephi.config.BotState;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,15 +10,15 @@ import java.util.Map;
 @Component
 public class UserDataCache implements DataCache {
 
-    private Map<Integer, BotState> usersBotStates = new HashMap<>();
+    private Map<Long, BotState> usersBotStates = new HashMap<>();
 
     @Override
-    public void setUsersCurrentBotState(int userId, BotState botState) {
+    public void setUsersCurrentBotState(Long userId, BotState botState) {
         usersBotStates.put(userId, botState);
     }
 
     @Override
-    public BotState getUsersCurrentBotState(int userId) {
+    public BotState getUsersCurrentBotState(Long userId) {
         BotState botState = usersBotStates.get(userId);
         if (botState == null) {
             botState = BotState.NEW_USER;
